requirements.txt
Flask==1.1.2
google-cloud-datastore==2.1.0
urllib3==1.25.10
requests==2.22.0
moviepy==1.0.1
python_craigslist==1.1.1
beautifulsoup4==4.9.3
craigslist==0.1.10
instabot==0.117.0
pandas==1.2.3
Pillow==8.2.0
tweepy==3.10.0
wget==3.2
